+++ js/setup.js
@@ -1,0 +1,23 @@
+const PHOTOS_COUNT = 25;
+
+const LikesCount = {
+  min: 15,
+  max: 200
+};
+
+const PhrasesCount = {
+  min: 1,
+  max: 2
+};
+
+const CommentsCount = {
+  min: 0,
+  max: 3
+};
+
+const AvatarCount = {
+  min: 1,
+  max: 6
+};
+
+export {PHOTOS_COUNT, LikesCount, PhrasesCount, CommentsCount, AvatarCount};
